Add CI using autopkgtest and travis.debian.net.
Right now this just makes sure that the example packages build at all,
to catch FTBFS problems like the 5.2 release. The next steps are to make
sure the example packages actually do the changes we expect them to do
and to make sure we have complete test coverage of all our
functionality.