Also make examples Multi-Arch: foreign
This doesn't matter for most config packages, but for the cases it does,
setting M-A: foreign is more often correct than not. In particular, an
Architecture: all package with M-A: no (or unset) cannot be used as a
build dependency during cross-builds, and requires that its dependencies
are installed by the native architecture, which unnecessarily
complicates things like in-place architecture changes.

There are a few cases where Arch: all and M-A: no is correct (e.g., if
maintainer scripts compile things for the native architecture on the
target machine), but just as the right default for our documentation is
Arch: all, the right default for our documentation is M-A: foreign. A
packager who's doing something where architectures are relevant should
set these fields appropriately.